GTSRB/requirements.txt
h5py
numpy
pandas
Pillow
tensorboard==1.15.0
torch==1.8.0
torchvision==0.9.0
scikit-image==0.17.2
scikit-learn
